+++ .env.example
@@ -1,4 +1,3 @@
 VITE_MEILISEARCH_HOST=https://search.bettergov.ph
 VITE_MEILISEARCH_PORT=443
 VITE_MEILISEARCH_SEARCH_API_KEY= # Meilisearch Search API Key
-VITE_MAILER_API=https://better-gov-mailer.onrender.com
