+++ package.json
@@ -30,20 +30,14 @@
     "fetch:contractor-profiles": "node ./scripts/fetch_contractor_profiles.cjs"
   },
   "dependencies": {
-    "@hookform/resolvers": "^5.2.2",
     "@icons-pack/react-simple-icons": "^13.7.0",
     "@meilisearch/instant-meilisearch": "^0.26.0",
-    "@radix-ui/react-dialog": "^1.1.15",
+    "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-scroll-area": "^1.0.5",
-    "@radix-ui/react-select": "^2.2.6",
-    "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-tabs": "^1.0.4",
-    "@radix-ui/react-tooltip": "^1.2.8",
     "@types/react-leaflet": "^3.0.0",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
+    "clsx": "^2.1.0",
     "csv-parser": "^3.2.0",
     "date-fns": "^3.3.1",
     "dotenv": "^16.5.0",
@@ -55,20 +49,16 @@
     "leaflet": "^1.9.4",
     "lucide-react": "^0.513.0",
     "meilisearch": "^0.50.0",
-    "next-themes": "^0.4.6",
     "nuqs": "^2.6.0",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-helmet-async": "^2.0.5",
-    "react-hook-form": "^7.63.0",
     "react-i18next": "^15.7.3",
     "react-instantsearch": "^7.15.8",
     "react-leaflet": "5.0.0-rc.2",
     "react-router-dom": "^6.22.2",
     "recharts": "^2.15.3",
-    "sonner": "^2.0.7",
-    "tailwind-merge": "^2.6.0",
-    "tailwindcss-animate": "^1.0.7",
+    "tailwind-merge": "^2.2.1",
     "wrangler": "^4.19.1",
     "zod": "^4.1.11"
   },
