--- src/components/ui/form/constant.tsx
@@ -1,23 +1,0 @@
-import * as React from 'react';
-import { type FieldPath, type FieldValues } from 'react-hook-form';
-
-type FormFieldContextValue<
-  TFieldValues extends FieldValues = FieldValues,
-  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
-> = {
-  name: TName;
-};
-
-const FormFieldContext = React.createContext<FormFieldContextValue>(
-  {} as FormFieldContextValue
-);
-
-type FormItemContextValue = {
-  id: string;
-};
-
-const FormItemContext = React.createContext<FormItemContextValue>(
-  {} as FormItemContextValue
-);
-
-export { FormFieldContext, FormItemContext };
